修改返回值

+++ src/main/java/info/hb/rpc/core/domain/QpsAndTps.java
@@ -3,33 +3,33 @@ package info.hb.rpc.core.domain;
 public class QpsAndTps {
 
 	// 每秒请求次数
-	private double qps;
+	private long qps;
 	// 每秒请求字节数
-	private double tps;
+	private long tps;
 
 	public QpsAndTps() {
 		super();
 	}
 
-	public QpsAndTps(double qps, double tps) {
+	public QpsAndTps(long qps, long tps) {
 		super();
 		this.qps = qps;
 		this.tps = tps;
 	}
 
-	public double getQps() {
+	public long getQps() {
 		return qps;
 	}
 
-	public void setQps(double qps) {
+	public void setQps(long qps) {
 		this.qps = qps;
 	}
 
-	public double getTps() {
+	public long getTps() {
 		return tps;
 	}
 
-	public void setTps(double tps) {
+	public void setTps(long tps) {
 		this.tps = tps;
 	}
